+++ api/app/controllers/DetallesCCController.php
@@ -2,24 +2,35 @@
 
 namespace app\controllers;
 
+use app\auth\GenerateJWT;
 use app\models\DetallesCC;
 use app\config\Entity;
 use app\config\Error;
 
-class DetallesCCController {
+class DetallesCCController
+{
     private $detallesCC;
     private $entity;
+    private $validationToken;
 
-    public function __construct() {
+    public function __construct()
+    {
         $this->detallesCC = new DetallesCC();
         $this->entity = new Entity(get_class($this->detallesCC));
+        $this->validationToken = new GenerateJWT();
     }
 
-    public function getDetallesCC(){
+    public function getDetallesCC($token)
+    {
+        if (!$this->validationToken->validateToken($token))
+            throw new Error(401, "Unauthorized request");
         return json_encode($this->entity->findAll());
     }
-    
-    public function getDetalleCCConId($id){
+
+    public function getDetalleCCConId($id, $token)
+    {
+        if (!$this->validationToken->validateToken($token))
+            throw new Error(401, "Unauthorized request");
         $data = $this->entity->findById($id);
         if (empty($data)) {
             $error = new Error(400, "Detalles del carrito no se encuentra");
@@ -29,11 +40,17 @@ class DetallesCCController {
         }
     }
 
-    public function createDetalleCC($entity){
+    public function createDetalleCC($entity, $token)
+    {
+        if (!$this->validationToken->validateToken($token))
+            throw new Error(401, "Unauthorized request");
         return json_encode($this->entity->save($entity));
     }
 
-    public function updateDetalleCC($id, $entity){
+    public function updateDetalleCC($id, $entity, $token)
+    {
+        if (!$this->validationToken->validateToken($token))
+            throw new Error(401, "Unauthorized request");
         $data = $this->entity->update($id, $entity);
         if (empty($data)) {
             $error = new Error(400, "Detalles del carrito no se encuentra");
@@ -43,7 +60,10 @@ class DetallesCCController {
         }
     }
 
-    public function deleteDetalleCC($id){
+    public function deleteDetalleCC($id, $token)
+    {
+        if (!$this->validationToken->validateToken($token))
+            throw new Error(401, "Unauthorized request");
         $data = $this->entity->delete($id);
         if (empty($data)) {
             $error = new Error(400, "Detalles del carrito no se encuentra");
@@ -52,7 +72,4 @@ class DetallesCCController {
             return json_encode($data);
         }
     }
-
 }
-
-?>
