+++ api/app/controllers/DetallesOController.php
@@ -2,24 +2,35 @@
 
 namespace app\controllers;
 
+use app\auth\GenerateJWT;
 use app\models\DetallesOrdenes;
 use app\config\Entity;
 use app\config\Error;
 
-class DetallesOController {
+class DetallesOController
+{
     private $detallesO;
     private $entity;
+    private $validationToken;
 
-    public function __construct() {
+    public function __construct()
+    {
         $this->detallesO = new DetallesOrdenes();
         $this->entity = new Entity(get_class($this->detallesO));
+        $this->validationToken = new GenerateJWT();
     }
 
-    public function getDetallesO(){
+    public function getDetallesO($token)
+    {
+        if (!$this->validationToken->validateToken($token))
+            throw new Error(401, "Unauthorized request");
         return json_encode($this->entity->findAll());
     }
-    
-    public function getDetalleOConId($id){
+
+    public function getDetalleOConId($id, $token)
+    {
+        if (!$this->validationToken->validateToken($token))
+            throw new Error(401, "Unauthorized request");
         $data = $this->entity->findById($id);
         if (empty($data)) {
             $error = new Error(400, "Detalle de orden no se encuentra");
@@ -29,11 +40,17 @@ class DetallesOController {
         }
     }
 
-    public function createDetalleO($entity){
+    public function createDetalleO($entity, $token)
+    {
+        if (!$this->validationToken->validateToken($token))
+            throw new Error(401, "Unauthorized request");
         return json_encode($this->entity->save($entity));
     }
 
-    public function updateDetalleO($id, $entity){
+    public function updateDetalleO($id, $entity, $token)
+    {
+        if (!$this->validationToken->validateToken($token))
+            throw new Error(401, "Unauthorized request");
         $data = $this->entity->update($id, $entity);
         if (empty($data)) {
             $error = new Error(400, "Detalle de orden no se encuentra");
@@ -43,7 +60,10 @@ class DetallesOController {
         }
     }
 
-    public function deleteDetalleO($id){
+    public function deleteDetalleO($id, $token)
+    {
+        if (!$this->validationToken->validateToken($token))
+            throw new Error(401, "Unauthorized request");
         $data = $this->entity->delete($id);
         if (empty($data)) {
             $error = new Error(400, "Detalle de orden no se encuentra");
@@ -52,7 +72,4 @@ class DetallesOController {
             return json_encode($data);
         }
     }
-
 }
-
-?>
