+++ api/app/controllers/ProductosController.php
@@ -2,24 +2,35 @@
 
 namespace app\controllers;
 
+use app\auth\GenerateJWT;
 use app\models\Productos;
 use app\config\Entity;
 use app\config\Error;
 
-class ProductosController {
+class ProductosController
+{
     private $productos;
     private $entity;
+    private $validationToken;
 
-    public function __construct() {
+    public function __construct()
+    {
         $this->productos = new Productos();
         $this->entity = new Entity(get_class($this->productos));
+        $this->validationToken = new GenerateJWT();
     }
 
-    public function getProductos(){
+    public function getProductos($token)
+    {
+        if (!$this->validationToken->validateToken($token))
+            throw new Error(401, "Unauthorized request");
         return json_encode($this->entity->findAll());
     }
-    
-    public function getProductoConId($id){
+
+    public function getProductoConId($id, $token)
+    {
+        if (!$this->validationToken->validateToken($token))
+            throw new Error(401, "Unauthorized request");
         $data = $this->entity->findById($id);
         if (empty($data)) {
             $error = new Error(400, "Producto no se encuentra");
@@ -29,11 +40,17 @@ class ProductosController {
         }
     }
 
-    public function createProducto($entity){
+    public function createProducto($entity, $token)
+    {
+        if (!$this->validationToken->validateToken($token))
+            throw new Error(401, "Unauthorized request");
         return json_encode($this->entity->save($entity));
     }
 
-    public function updateProducto($id, $entity){
+    public function updateProducto($id, $entity, $token)
+    {
+        if (!$this->validationToken->validateToken($token))
+            throw new Error(401, "Unauthorized request");
         $data = $this->entity->update($id, $entity);
         if (empty($data)) {
             $error = new Error(400, "Producto no se encuentra");
@@ -43,7 +60,10 @@ class ProductosController {
         }
     }
 
-    public function deleteProducto($id){
+    public function deleteProducto($id, $token)
+    {
+        if (!$this->validationToken->validateToken($token))
+            throw new Error(401, "Unauthorized request");
         $data = $this->entity->delete($id);
         if (empty($data)) {
             $error = new Error(400, "Producto no se encuentra");
@@ -52,7 +72,4 @@ class ProductosController {
             return json_encode($data);
         }
     }
-
 }
-
-?>
