+++ api/app/controllers/RolesController.php
@@ -2,6 +2,7 @@
 
 namespace app\controllers;
 
+use app\auth\GenerateJWT;
 use app\models\Roles;
 use app\config\Entity;
 use app\config\Error;
@@ -10,20 +11,26 @@ class RolesController
 {
     private $roles;
     private $entity;
+    private $validationToken;
 
     public function __construct()
     {
         $this->roles = new Roles();
         $this->entity = new Entity(get_class($this->roles));
+        $this->validationToken = new GenerateJWT();
     }
 
-    public function getRoles()
+    public function getRoles($token)
     {
+        if (!$this->validationToken->validateToken($token))
+            throw new Error(401, "Unauthorized request");
         return json_encode($this->entity->findAll());
     }
 
-    public function getRolConId($id)
+    public function getRolConId($id, $token)
     {
+        if (!$this->validationToken->validateToken($token))
+            throw new Error(401, "Unauthorized request");
         $data = $this->entity->findById($id);
         if (empty($data)) {
             $error = new Error(400, "Rol no se encuentra");
@@ -33,13 +40,17 @@ class RolesController
         }
     }
 
-    public function createRol($entity)
+    public function createRol($entity, $token)
     {
+        if (!$this->validationToken->validateToken($token))
+            throw new Error(401, "Unauthorized request");
         return json_encode($this->entity->save($entity));
     }
 
-    public function updateRol($id, $entity)
+    public function updateRol($id, $entity, $token)
     {
+        if (!$this->validationToken->validateToken($token))
+            throw new Error(401, "Unauthorized request");
         $data = $this->entity->update($id, $entity);
         if (empty($data)) {
             $error = new Error(400, "Rol no se encuentra");
@@ -49,8 +60,10 @@ class RolesController
         }
     }
 
-    public function deleteRol($id)
+    public function deleteRol($id, $token)
     {
+        if (!$this->validationToken->validateToken($token))
+            throw new Error(401, "Unauthorized request");
         $data = $this->entity->delete($id);
         if (empty($data)) {
             $error = new Error(400, "Rol no se encuentra");
